refactor(tests): move public-API db tests from inline mod to tests/db_*.rs
Migrates 20 tests out of `src/core/db.rs::mod tests` into 5 per-feature
integration test files under `tests/`, leaving 23 white-box tests inline
with a module-level annotation listing the private items they exercise.
Pure mechanical refactor: no production code, visibility, or test logic
changes — only test relocation with adjusted `use` paths.

New files (each ≤ 1500-LOC budget):
- tests/db_open.rs (6 tests) — open lifecycle + compact
- tests/db_schema_v9.rs (5 tests) — fresh v9 schema/constraints/defaults
- tests/db_schema_migrations.rs (2 tests) — DDL repair + content-hash backfill
- tests/db_quaid_config.rs (5 tests) — QuaidConfig roundtrips via pub API
- tests/db_model_channel.rs (2 tests, online-model gated) — small/large model

Inline residue (23 tests, 493 LOC, under the < 500 LOC target):
private helpers exercised include `open_connection`,
`preflight_existing_schema`, `table_exists`,
`pages_needs_namespace_unique_rebuild`, `recover_crash_partial_fresh_db`,
`read_bootstrap_registry_model`, `model_alias_for_model_id`,
`ensure_embedding_model_registry`, `persist_model_metadata`,
`is_bootstrap_fresh_db`, `format_model_mismatch`, and
`QuaidConfig::{from_model, to_model_config}`.

Verification:
  baseline (parent commit a1018eb): passed=1357 failed=0
  this commit:                       passed=1357 failed=0

Both clippy channels green:
  cargo clippy --all-targets --locked -- -D warnings
  cargo clippy --all-targets --no-default-features \
    --features bundled,online-model --locked -- -D warnings

src/core/db.rs:  2040 LOC -> 1479 LOC (-561 LOC, -27%)

Part of OpenSpec change `extract-inline-tests-to-integration`.
