+++ src/core/db.rs
@@ -975,6 +975,15 @@ pub fn set_version(conn: &Connection) -> Result<(), DbError> {
     Ok(())
 }
 
+// reason: white-box; needs `seed_existing_db`, `open_connection`,
+// `preflight_existing_schema`, `table_exists`, `pages_needs_namespace_unique_rebuild`,
+// `recover_crash_partial_fresh_db`, `read_bootstrap_registry_model`,
+// `model_alias_for_model_id`, `ensure_embedding_model_registry`,
+// `persist_model_metadata`, `is_bootstrap_fresh_db`, `format_model_mismatch`,
+// and the private `QuaidConfig::{from_model, to_model_config}` methods.
+// Public-API tests for `open`, `open_with_model`, `init`, `compact`,
+// `read_quaid_config`, `write_quaid_config`, and `QuaidConfig` round-trips
+// live under `tests/db_*.rs`.
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1009,450 +1018,6 @@ mod tests {
             [schema_version.to_string()],
         )
         .unwrap();
-    }
-
-    #[test]
-    fn open_creates_all_expected_tables() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let conn = open(db_path.to_str().unwrap()).unwrap();
-
-        let tables: Vec<String> = conn
-            .prepare(
-                "SELECT name FROM sqlite_master \
-                 WHERE type = 'table' AND name NOT LIKE 'sqlite_%' \
-                 ORDER BY name",
-            )
-            .unwrap()
-            .query_map([], |row| row.get(0))
-            .unwrap()
-            .filter_map(Result::ok)
-            .collect();
-
-        let expected = [
-            "assertions",
-            "quaid_config",
-            "collections",
-            "collection_owners",
-            "config",
-            "contradictions",
-            "correction_sessions",
-            "embedding_jobs",
-            "embedding_models",
-            "extraction_queue",
-            "file_state",
-            "import_manifest",
-            "knowledge_gaps",
-            "links",
-            "namespaces",
-            "page_embeddings",
-            "page_fts",
-            "pages",
-            "raw_data",
-            "raw_imports",
-            "serve_sessions",
-            "tags",
-            "timeline_entries",
-        ];
-
-        for name in &expected {
-            assert!(
-                tables.contains(&(*name).to_string()),
-                "missing table: {name}"
-            );
-        }
-    }
-
-    #[test]
-    fn open_sets_user_version_to_9() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let conn = open(db_path.to_str().unwrap()).unwrap();
-
-        let version: i64 = conn
-            .query_row("PRAGMA user_version", [], |row| row.get(0))
-            .unwrap();
-        assert_eq!(version, 9);
-    }
-
-    #[test]
-    fn fresh_v9_schema_includes_conversation_memory_artifacts_and_defaults() {
-        let conn = open(":memory:").unwrap();
-
-        let page_columns: Vec<String> = conn
-            .prepare("PRAGMA table_info(pages)")
-            .unwrap()
-            .query_map([], |row| row.get(1))
-            .unwrap()
-            .collect::<Result<_, _>>()
-            .unwrap();
-        assert!(page_columns.contains(&"superseded_by".to_string()));
-
-        let page_foreign_keys: Vec<(String, String)> = conn
-            .prepare("PRAGMA foreign_key_list(pages)")
-            .unwrap()
-            .query_map([], |row| Ok((row.get(2)?, row.get(3)?)))
-            .unwrap()
-            .collect::<Result<_, _>>()
-            .unwrap();
-        assert!(page_foreign_keys.contains(&("pages".to_string(), "superseded_by".to_string())));
-
-        let supersede_index_sql: String = conn
-            .query_row(
-                "SELECT sql FROM sqlite_master WHERE type = 'index' AND name = 'idx_pages_supersede_head'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert!(supersede_index_sql.contains("WHERE superseded_by IS NULL"));
-
-        let session_index_sql: String = conn
-            .query_row(
-                "SELECT sql FROM sqlite_master WHERE type = 'index' AND name = 'idx_pages_session'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert!(session_index_sql.contains("json_valid(frontmatter)"));
-        assert!(session_index_sql.contains("$.session_id"));
-
-        let queue_table_sql: String = conn
-            .query_row(
-                "SELECT sql FROM sqlite_master WHERE type = 'table' AND name = 'extraction_queue'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert!(queue_table_sql.contains("trigger_kind IN ('debounce', 'session_close', 'manual')"));
-        assert!(queue_table_sql.contains("status IN ('pending', 'running', 'done', 'failed')"));
-
-        let pending_index_sql: String = conn
-            .query_row(
-                "SELECT sql FROM sqlite_master WHERE type = 'index' AND name = 'idx_extraction_queue_pending'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert!(pending_index_sql.contains("WHERE status = 'pending'"));
-
-        let correction_table_sql: String = conn
-            .query_row(
-                "SELECT sql FROM sqlite_master WHERE type = 'table' AND name = 'correction_sessions'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert!(correction_table_sql
-            .contains("status IN ('open', 'committed', 'abandoned', 'expired')"));
-        assert!(correction_table_sql.contains("json_valid(exchange_log)"));
-
-        let correction_index_sql: String = conn
-            .query_row(
-                "SELECT sql FROM sqlite_master WHERE type = 'index' AND name = 'idx_correction_open'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert!(correction_index_sql.contains("WHERE status = 'open'"));
-
-        let config_rows: Vec<(String, String)> = conn
-            .prepare(
-                "SELECT key, value FROM config
-                 WHERE key IN (
-                     'version',
-                     'memory.location',
-                     'corrections.history_on_disk',
-                     'extraction.max_retries',
-                     'extraction.enabled',
-                     'extraction.model_alias',
-                     'extraction.window_turns',
-                     'extraction.debounce_ms',
-                     'extraction.idle_close_ms',
-                     'extraction.retention_days',
-                     'fact_resolution.dedup_cosine_min',
-                     'fact_resolution.supersede_cosine_min'
-                 )
-                 ORDER BY key",
-            )
-            .unwrap()
-            .query_map([], |row| Ok((row.get(0)?, row.get(1)?)))
-            .unwrap()
-            .collect::<Result<_, _>>()
-            .unwrap();
-        assert_eq!(
-            config_rows,
-            vec![
-                (
-                    "corrections.history_on_disk".to_string(),
-                    "false".to_string()
-                ),
-                ("extraction.debounce_ms".to_string(), "5000".to_string()),
-                ("extraction.enabled".to_string(), "false".to_string()),
-                ("extraction.idle_close_ms".to_string(), "60000".to_string()),
-                ("extraction.max_retries".to_string(), "3".to_string()),
-                (
-                    "extraction.model_alias".to_string(),
-                    "phi-3.5-mini".to_string()
-                ),
-                ("extraction.retention_days".to_string(), "30".to_string()),
-                ("extraction.window_turns".to_string(), "5".to_string()),
-                (
-                    "fact_resolution.dedup_cosine_min".to_string(),
-                    "0.92".to_string()
-                ),
-                (
-                    "fact_resolution.supersede_cosine_min".to_string(),
-                    "0.4".to_string()
-                ),
-                ("memory.location".to_string(), "vault-subdir".to_string()),
-                ("version".to_string(), "9".to_string()),
-            ]
-        );
-    }
-
-    #[test]
-    fn fresh_v9_schema_enforces_superseded_by_foreign_key() {
-        let conn = open(":memory:").unwrap();
-
-        let err = conn
-            .execute(
-                "INSERT INTO pages
-                     (slug, uuid, type, title, summary, compiled_truth, timeline, frontmatter, wing, room, superseded_by, version)
-                 VALUES ('notes/invalid-supersede', ?1, 'concept', 'Invalid', '', '', '', '{}', 'notes', '', 9999, 1)",
-                [uuid::Uuid::now_v7().to_string()],
-            )
-            .expect_err("invalid superseded_by should fail");
-
-        assert!(matches!(err, rusqlite::Error::SqliteFailure(_, _)));
-    }
-
-    #[test]
-    fn fresh_v9_schema_rejects_invalid_extraction_queue_trigger_kind() {
-        let conn = open(":memory:").unwrap();
-
-        let err = conn
-            .execute(
-                "INSERT INTO extraction_queue
-                     (session_id, conversation_path, trigger_kind, enqueued_at, scheduled_for, status)
-                 VALUES ('s1', 'conversations/2026-05-04/s1.md', 'arbitrary', '2026-05-04T00:00:00Z', '2026-05-04T00:00:05Z', 'pending')",
-                [],
-            )
-            .expect_err("invalid trigger_kind should fail");
-
-        assert!(matches!(err, rusqlite::Error::SqliteFailure(_, _)));
-    }
-
-    #[test]
-    fn fresh_v9_schema_rejects_invalid_extraction_queue_status() {
-        let conn = open(":memory:").unwrap();
-
-        let err = conn
-            .execute(
-                "INSERT INTO extraction_queue
-                     (session_id, conversation_path, trigger_kind, enqueued_at, scheduled_for, status)
-                 VALUES ('s1', 'conversations/2026-05-04/s1.md', 'debounce', '2026-05-04T00:00:00Z', '2026-05-04T00:00:05Z', 'queued')",
-                [],
-            )
-            .expect_err("invalid status should fail");
-
-        assert!(matches!(err, rusqlite::Error::SqliteFailure(_, _)));
-    }
-
-    #[test]
-    fn fresh_v9_schema_rejects_invalid_correction_session_status() {
-        let conn = open(":memory:").unwrap();
-
-        let err = conn
-            .execute(
-                "INSERT INTO correction_sessions
-                     (correction_id, fact_slug, exchange_log, turns_used, status, created_at, expires_at)
-                 VALUES (?1, 'facts/example', '[]', 0, 'paused', '2026-05-05T00:00:00Z', '2026-05-05T01:00:00Z')",
-                [uuid::Uuid::now_v7().to_string()],
-            )
-            .expect_err("invalid correction session status should fail");
-
-        assert!(matches!(err, rusqlite::Error::SqliteFailure(_, _)));
-    }
-
-    #[test]
-    fn open_enables_wal_and_foreign_keys() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let conn = open(db_path.to_str().unwrap()).unwrap();
-
-        let journal: String = conn
-            .query_row("PRAGMA journal_mode", [], |row| row.get(0))
-            .unwrap();
-        assert_eq!(journal.to_lowercase(), "wal");
-
-        let fk: i64 = conn
-            .query_row("PRAGMA foreign_keys", [], |row| row.get(0))
-            .unwrap();
-        assert_eq!(fk, 1);
-    }
-
-    #[test]
-    fn open_rejects_nonexistent_parent_dir() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let missing = dir.path().join("missing-parent").join("memory.db");
-        let result = open(missing.to_str().unwrap());
-        assert!(result.is_err());
-        assert!(matches!(result.unwrap_err(), DbError::PathNotFound { .. }));
-    }
-
-    #[test]
-    fn open_is_idempotent() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let path_str = db_path.to_str().unwrap();
-
-        let conn1 = open(path_str).unwrap();
-        drop(conn1);
-
-        let conn2 = open(path_str).unwrap();
-        let version: i64 = conn2
-            .query_row("PRAGMA user_version", [], |row| row.get(0))
-            .unwrap();
-        assert_eq!(version, 9);
-    }
-
-    #[test]
-    fn open_replaces_buggy_pages_update_trigger_for_quarantined_rows() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let path_str = db_path.to_str().unwrap();
-
-        let conn = open(path_str).unwrap();
-        conn.execute_batch(
-            "DROP TRIGGER IF EXISTS pages_au;
-             CREATE TRIGGER pages_au AFTER UPDATE ON pages BEGIN
-                 INSERT INTO page_fts(page_fts, rowid, title, slug, compiled_truth, timeline)
-                 VALUES ('delete', old.id, old.title, old.slug, old.compiled_truth, old.timeline);
-                 INSERT INTO page_fts(rowid, title, slug, compiled_truth, timeline)
-                 SELECT new.id, new.title, new.slug, new.compiled_truth, new.timeline
-                 WHERE new.quarantined_at IS NULL;
-             END;",
-        )
-        .unwrap();
-        drop(conn);
-
-        let conn = open(path_str).unwrap();
-        let collection_id: i64 = conn
-            .query_row(
-                "SELECT id FROM collections ORDER BY id LIMIT 1",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        conn.execute(
-            "INSERT INTO pages
-                 (collection_id, slug, uuid, type, title, summary, compiled_truth, timeline, frontmatter, wing, room, version)
-             VALUES (?1, 'notes/quarantined', ?2, 'concept', 'Quarantined', '', 'before restore', '', '{}', 'notes', '', 1)",
-            params![collection_id, uuid::Uuid::now_v7().to_string()],
-        )
-        .unwrap();
-        let page_id = conn.last_insert_rowid();
-
-        conn.execute(
-            "UPDATE pages
-             SET quarantined_at = '2026-04-25T00:00:00Z'
-             WHERE id = ?1",
-            [page_id],
-        )
-        .unwrap();
-
-        conn.execute(
-            "UPDATE pages
-             SET slug = 'notes/restored',
-                 title = 'Restored',
-                 compiled_truth = 'after restore',
-                 quarantined_at = NULL
-             WHERE id = ?1",
-            [page_id],
-        )
-        .unwrap();
-
-        let count: i64 = conn
-            .query_row(
-                "SELECT COUNT(*)
-                 FROM page_fts
-                 WHERE page_fts MATCH 'after'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(count, 1);
-    }
-
-    #[test]
-    fn open_backfills_raw_import_content_hash_for_existing_rows() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let path_str = db_path.to_str().unwrap();
-
-        let conn = open(path_str).unwrap();
-        conn.execute(
-            "INSERT INTO pages (slug, uuid, type, title)
-             VALUES ('notes/hash-test', ?1, 'concept', 'notes/hash-test')",
-            [crate::core::page_uuid::generate_uuid_v7()],
-        )
-        .unwrap();
-        let page_id: i64 = conn
-            .query_row(
-                "SELECT id FROM pages WHERE slug = 'notes/hash-test'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        conn.execute(
-            "INSERT INTO raw_imports (page_id, import_id, is_active, content_hash, raw_bytes, file_path)
-             VALUES (?1, 'import-1', 1, ?2, ?3, 'notes/hash-test.md')",
-            params![
-                page_id,
-                crate::core::raw_imports::content_hash_hex(b"hello"),
-                b"hello"
-            ],
-        )
-        .unwrap();
-        drop(conn);
-
-        let conn = Connection::open(path_str).unwrap();
-        conn.execute_batch(
-            "DROP INDEX IF EXISTS idx_raw_imports_content_hash;
-             ALTER TABLE raw_imports RENAME TO raw_imports_old;
-             CREATE TABLE raw_imports (
-                 id         INTEGER PRIMARY KEY AUTOINCREMENT,
-                 page_id    INTEGER NOT NULL REFERENCES pages(id) ON DELETE CASCADE,
-                 import_id  TEXT    NOT NULL,
-                 is_active  INTEGER NOT NULL DEFAULT 1,
-                 raw_bytes  BLOB    NOT NULL,
-                 file_path  TEXT    NOT NULL,
-                 created_at TEXT    NOT NULL DEFAULT (strftime('%Y-%m-%dT%H:%M:%SZ', 'now')),
-                 UNIQUE(page_id, import_id)
-             );
-             INSERT INTO raw_imports (id, page_id, import_id, is_active, raw_bytes, file_path, created_at)
-             SELECT id, page_id, import_id, is_active, raw_bytes, file_path, created_at
-             FROM raw_imports_old;
-             DROP TABLE raw_imports_old;
-             CREATE INDEX idx_raw_imports_page ON raw_imports(page_id);
-             CREATE INDEX idx_raw_imports_active ON raw_imports(page_id, is_active)
-                 WHERE is_active = 1;",
-        )
-        .unwrap();
-        drop(conn);
-
-        let conn = open(path_str).unwrap();
-        let content_hash: String = conn
-            .query_row(
-                "SELECT content_hash FROM raw_imports WHERE page_id = ?1",
-                [page_id],
-                |row| row.get(0),
-            )
-            .unwrap();
-        assert_eq!(
-            content_hash,
-            crate::core::raw_imports::content_hash_hex(b"hello")
-        );
     }
 
     #[test]
@@ -1744,34 +1309,6 @@ mod tests {
     }
 
     #[test]
-    fn compact_checkpoints_wal() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let conn = open(db_path.to_str().unwrap()).unwrap();
-        assert!(compact(&conn).is_ok());
-    }
-
-    #[test]
-    fn open_seeds_default_embedding_model() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let opened = open_with_model(db_path.to_str().unwrap(), &default_model()).unwrap();
-
-        let (name, dims, active): (String, i64, i64) = opened
-            .conn
-            .query_row(
-                "SELECT name, dimensions, active FROM embedding_models WHERE active = 1",
-                [],
-                |row| Ok((row.get(0)?, row.get(1)?, row.get(2)?)),
-            )
-            .unwrap();
-
-        assert_eq!(name, "BAAI/bge-small-en-v1.5");
-        assert_eq!(dims, 384);
-        assert_eq!(active, 1);
-    }
-
-    #[test]
     fn persist_model_metadata_writes_registry_quaid_config_and_legacy_config() {
         let conn = open_connection(":memory:").unwrap();
 
@@ -1861,47 +1398,6 @@ mod tests {
     }
 
     #[test]
-    fn quaid_config_roundtrip_preserves_values() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let opened = open_with_model(db_path.to_str().unwrap(), &default_model()).unwrap();
-
-        let config = read_quaid_config(&opened.conn).unwrap().unwrap();
-        assert_eq!(
-            config,
-            QuaidConfig {
-                model_id: "BAAI/bge-small-en-v1.5".to_owned(),
-                model_alias: "small".to_owned(),
-                embedding_dim: 384,
-                schema_version: 9,
-            }
-        );
-    }
-
-    #[test]
-    fn empty_quaid_config_reads_as_missing() {
-        let conn = open(":memory:").unwrap();
-        conn.execute("DELETE FROM quaid_config", []).unwrap();
-
-        let config = read_quaid_config(&conn).unwrap();
-        assert!(config.is_none());
-    }
-
-    #[test]
-    fn incomplete_quaid_config_returns_schema_error() {
-        let conn = open(":memory:").unwrap();
-        conn.execute("DELETE FROM quaid_config", []).unwrap();
-        conn.execute(
-            "INSERT INTO quaid_config (key, value) VALUES ('model_id', 'BAAI/bge-small-en-v1.5')",
-            [],
-        )
-        .unwrap();
-
-        let err = read_quaid_config(&conn).unwrap_err();
-        assert!(matches!(err, DbError::Schema { .. }));
-    }
-
-    #[test]
     fn read_quaid_config_rejects_non_integer_embedding_dimensions() {
         let conn = open_connection(":memory:").unwrap();
         write_quaid_config(
@@ -1950,35 +1446,6 @@ mod tests {
     }
 
     #[test]
-    fn missing_quaid_config_requires_reinit_once_pages_exist() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test_memory.db");
-        let opened = open_with_model(db_path.to_str().unwrap(), &default_model()).unwrap();
-        let collection_id: i64 = opened
-            .conn
-            .query_row(
-                "SELECT id FROM collections WHERE name = 'default'",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap();
-        opened
-            .conn
-            .execute(
-                "INSERT INTO pages
-                     (collection_id, slug, uuid, type, title, summary, compiled_truth, timeline, frontmatter, wing, room, version)
-                 VALUES (?1, 'notes/live', ?2, 'concept', 'Live', '', 'truth', '', '{}', 'notes', '', 1)",
-                params![collection_id, uuid::Uuid::now_v7().to_string()],
-            )
-            .unwrap();
-        opened.conn.execute("DELETE FROM quaid_config", []).unwrap();
-        drop(opened);
-
-        let reopened = open_with_model(db_path.to_str().unwrap(), &default_model());
-        assert!(matches!(reopened, Err(DbError::Schema { .. })));
-    }
-
-    #[test]
     fn mismatch_detection_returns_model_mismatch_error() {
         let stored = QuaidConfig {
             model_id: "BAAI/bge-small-en-v1.5".to_owned(),
@@ -2008,33 +1475,5 @@ mod tests {
         let message = format_model_mismatch(&stored, &requested, "memory.db");
 
         assert!(message.contains("QUAID_MODEL=org/custom-model quaid <command>"));
-    }
-
-    #[cfg(feature = "online-model")]
-    #[test]
-    fn init_with_small_then_open_with_large_errors() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("memory.db");
-
-        init(db_path.to_str().unwrap(), &resolve_model("small")).unwrap();
-        let err = open_with_model(db_path.to_str().unwrap(), &resolve_model("large")).unwrap_err();
-
-        assert!(matches!(err, DbError::ModelMismatch { .. }));
-    }
-
-    #[cfg(feature = "online-model")]
-    #[test]
-    fn init_with_large_then_open_with_large_succeeds() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("memory.db");
-
-        let init_opened =
-            open_with_model(db_path.to_str().unwrap(), &resolve_model("large")).unwrap();
-        drop(init_opened);
-
-        let reopened = open_with_model(db_path.to_str().unwrap(), &resolve_model("large")).unwrap();
-        let stored = read_quaid_config(&reopened.conn).unwrap().unwrap();
-        assert_eq!(stored.model_alias, "large");
-        assert_eq!(stored.embedding_dim, 1024);
     }
 }
